Create index.html
For the uninitiated, the Ranked mode in Fortnite is identical to the usual game modes (Battle Royale, Zero Build) but made for those who want to be in competitive lobbies. In each game, players compete for higher ranks that dictate their competitive skill level and the quality of players they compete against.